// src/types/benefit.ts
export interface CouponData {
  couponId: string;
  smartDutyFreeAppUseYn: string;
  offlineStoreUseYn: string;
  couponNameKo: string;
  couponNameEn: string;
  couponNameJp: string | null;
  couponNameCn: string | null;
  couponNameTw: string | null;
  couponType: string;
  couponTypeName: string;
  issuanceType: string;
  issuanceTypeName: string;
  discountApplicationType: string;
  discountApplicationTypeName: string;
  membershipBenefitType: string;
  membershipBenefitTypeName: string;
  targetProduct: string | null;
  issuanceStart: string;
  issuanceEnd: string;
  issuanceCycle: string;
  issuanceCycleName: string;
  usagePeriodCode: string;
  usagePeriodCodeName: string;
  usageStart: string | null;
  usageEnd: string | null;
  samePeriodYn: string | null;
  validityPeriodDays: number;
  airportCorpShareRate: number;
  dutyFreeBusinessShareRate: number;
  gradeDifferentiationBenefit: string | null;
  gradeDifferentiationBenefitName: string | null;
  displayYn: string;
  couponNotice: string;
  delYn: string;
  forceTerminateYn: string;
  forceTerminateReason: string | null;
  forceTerminateDt: string | null;
  issueStatus: string;
  issueDate: string;
  issueUser: string;
  regDate: string;
  regUser: string;
  updDate: string;
  updUser: string;
  issuanceConditionCode: string | null;
  issuanceConditionCodeName: string | null;
  issuanceAmountConditionYn: string | null;
  issuanceMinimumAmount: number | null;
  usageCountPerFlight: number | null;
  usageConditionCode: string | null;
  usageConditionCodeName: string | null;
  usageAmountConditionYn: string | null;
  usageMinimumAmount: number | null;
}